[Feature][James] tweaked responsiveness and padding
